Revert component ports (separated to another PR)
